Beers interpolation, using ordinary or modified method.

Create 4 interpolated points between each pair of given values. With the ordinary method, the given data points given will be included unchanged in the interpolated list. With the modified method, some smoothing occurs, and only the first and last points given are guaranteed to appear unchanged in the interpolated list.

beers_int_ordinary(points)

beers_int_modified(points)

Arguments

points
A list of at least 6 numbers, eg, populations over time.

Value

A list with four extra interpolated points between each pair of given points.

Examples

# Interpolate population of UK (1950, 1955, 1960, 1965, 1970, 1975) to yearly points

beers_int_ordinary(c(50616014, 51123707, 52433157, 54303107, 55634935, 56211947))

#>  [1] 50616014 50621417 50686057 50797277 50945395 51123707 51328658 51559670
#>  [9] 51818620 52108970 52433157 52789813 53171327 53562802 53945015 54303107
#> [17] 54629547 54923085 55186314 55422893 55634935 55822137 55981259 56105949
#> [25] 56186920 56211947

beers_int_modified(c(50616014, 51123707, 52433157, 54303107, 55634935, 56211947))

#>  [1] 50616014 50648230 50715139 50816709 50952909 51123707 51328897 51567841
#>  [9] 51839471 52141505 52470024 52819465 53182550 53550545 53914313 54265184
#> [17] 54595990 54901323 55177440 55422263 55634935 55815041 55962600 56077808
#> [25] 56160859 56211947