Beers subdivision, using ordinary or modified method.

With the ordinary method, each set of 5 subdivided values will always sum to the original data point given. With the modified method, some smoothing occurs, and this property is only true for the first and last given data point.

beers_sub_ordinary(points)

beers_sub_modified(points)

Arguments

points
A list of at least 5 numbers, eg, populations by 5-year age-band.

Value

For subdivision: a list 5 times as long as the original, with each point subdivided into 5.

Examples

# Subdivide population of UK (2015), ages 0-4, 5-9, 10-14, 15-19, 20-24 into single years.

beers_sub_ordinary(c(4042918, 3927745, 3529200, 3779712, 4174572))

#>  [1] 813537.0 807664.3 806238.0 807157.4 808321.3 807485.2 802548.0 791840.0
#>  [9] 774409.8 751462.0 726500.5 705041.0 693461.9 695698.0 708498.7 726170.2
#> [17] 743881.0 758811.6 770441.8 780407.3 791062.3 805191.9 825725.3 855447.8
#> [25] 897144.7

beers_sub_modified(c(4042918, 3927745, 3529200, 3779712, 4174572))

#>  [1] 798236.6 807514.6 812688.1 813757.1 810721.7 803725.5 793149.3 779459.2
#>  [9] 763746.8 747569.8 732528.8 720403.0 712863.2 710803.3 714398.6 722876.8
#> [17] 735142.7 750066.4 766475.2 783448.4 800733.9 817993.6 835083.8 852004.7
#> [25] 868756.1